+++ src/job.rs
@@ -241,7 +241,7 @@ pub(crate) async fn perform_job(context: &Context, mut connection: Connection<'_
                 info!(
                     context,
                     "job #{} not succeeded on try #{}, retry in {} seconds.",
-                    job.job_id as u32,
+                    job.job_id,
                     tries,
                     time_offset
                 );
